+++ docs/changelog.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Fix sorted set repository
+
 ### 0.22.6 - 2020-09-01
 
  - Improve Service.get_many performance